support amp in forward and backward
