+++ internlm/core/__init__.py
@@ -1,8 +1,11 @@
 from .engine import Engine
-from .naive_amp import NaiveAMPModel
+# from .naive_amp import NaiveAMPModel
+from .amp import convert_to_amp
+from .amp.naive_amp import NaiveAMPModel
 from .trainer import Trainer
 
 __all__ = [
+    "convert_to_amp",
     "NaiveAMPModel",
     "Engine",
     "Trainer",
